metric options for /grant-access and /sign-transaction added in metrics/view
